Updates to the 3.1 branch, based on previous updates to the trunk:
Added 209 from Trunk to 3.1 branch:
Fixed a bug that prevented colorpicker elements from being created (datatype was not being set).  Thank you to arjun for the bug report.

Added 215 too:
Fixed a bug where dynamic defaults for textboxes would override existing values in elements that are flagged as "include as hidden for users who can't see them"
Added awareness of user ids to the import process for FULLNAME and USERNAME lists

Added 217:
Fixed a bug where one level of permission checking at was failing for screens.
Made it so that users without view_form permission can be redirected back to the form after logging in.

Added this from 218:
Made it possible to disable colorpick elements.php
Fixed a bug in saving selectbox/checkbox/radiobutton elements if magic quotes is on (who is still using magic quotes?)
Fixed a bug in the saving of yn elements (datatype was not being set)
Allowed using element data handles instead of ids in the displayElement function, and many other functions, especially in the data class.
Added handling of submitted linked selectbox values when the values are blank/empty.

Added this from 219:
Fixed a reference to element IDs in LOE screen help text that should have been element data handles.
Made it so that when you specify a custom mail template, it can include or not include the .tpl part of the filename (files must still have that extension, but it doesn't matter if you type it into the notification options box or not).

Added this from 222:
Improved handling of default values for textboxes, so that any property of the user object, or the extended profile module in XOOPS 2.3, can be used in { } brackets.  Also improved handling of {DATE} and {TODAY} in textbox defaults, so they can be used multiple times and in conjunction with other { } values.
Fixed an obscure bug related to the internal conversion of element handles to element ids, when framework handles were also used on the same pageload.
Fixed an HTML bug in the rendering of hidden search boxes.
Fixed a bug in the application of filters to linked selectboxes when the filter is being done on a yes/no field ('yes' and 'no' were not being recognized as valid filter values!)

Addded 226:
Fixed a bug that could cause the 3.1 patch process for making database fields Nullable to fail.  Certain versions of MySQL report Null = NO when asked, and others report nothing, so the test for whether to patch a field needs to be if NULL != "YES" rather than NULL == "NO".

Added 227:
Improving the determination of elements used in derived value formulas when frameworks may, or may not, be in effect.

Added 228:
Tidying up the behaviour of derived value formulas at the time they are parsed.  A single error in one formula in a form would prevent any formulas from running, and now that is not true, and formulas where the elements that are being referenced cannot be identified, they have "Syntax Error" returned as their value instead.


git-svn-id: svn+ssh://svn.code.sf.net/p/formulize/code@229 e5624274-254a-0410-876d-c0e85fc73166