3.12 release:
From 232:
Fixing a bug in the new mechanism for determining the default values for textboxes.

From 235:
Fixed a bug in the determination of column headings, and data handles for import templates

From 236:
Fixed a bug in the creation of data tables that would ignore undisplayed fields that still need a container in the database.
Adjusted the security check when displaying forms so that it happens after the "single entry" determination is made.
Fixed a bug in readelements.php that caused blank defaults in subforms to be saved to the database even when they were left blank, in some situations.

From 237:
Fixed a bug where the creation of derived value fields during the update to the latest version, would fail.  Datatype was not being passed in when elements were inserted during the patch routine.

From 239:
Streamlined the basic security check that happens in initialize.php, to fix a bug that caused requests certain requests to end up in an endless loop (thanks Mark Boyden for the report!)

From 241:
Added fix for reported issue in the save.php file.

From 251:
Fixing the ordering of values in the quickFilter boxes.

From 252:
Added awareness of sorting and limits to "tableform" queries, which fixes several bug situations.

From 253:
Added awareness of pagination to tableform queries.

From 255:
HTML bug fixed in entriesdisplay.php that caused an extra closing div tag if the scrollbox was enabled on the page, but no entries were found/displayed.

From 259:
Fixed a bug that could cause the master $fid value in initialize.php to be set inaccurately, which had the bad side effect of making it impossible to return to the mainform when viewing a subform entry.

From 262:
Fixed a bug where the username fields would not export as names, but as ids instead.

From 263:
Made quickFilter boxes aware of the uitext of options.

From 265:
Fixing bugs related to apostrophes in form names, which can affect the computation of derived values.

From 266:
Improved handling of element handles when specifying the element with a handle in the displayElement function.
Made getCurrentURL function aware of the active domain based on the current HTTP_HOST setting in $_SERVER instead of the XOOPS URL, since Drupal multisite and other situations can make the host and URL different.

From 267:
Fixed a bug where names were not being displayed as grouping text in calculations, if the grouping was done with a usernames and fullnames list.

From 268:
Fixing a bug in the last commit...percentage breakdown calculations would break under certain (common) grouping situations.

From 269:
Fixed a bug where intro and thanks text for multipage screens was being mangled by the XOOPS/ICMS cores.
Made multipage screens use their title as the titlebar text for the form, instead of the form name.
Fixed a bug where selected options that have apostrophes were not being recognized as the saved value in the database.
Fixed an obscure bug where default values in radio buttons were being ignored if the entry was saved without any value for the radio buttons.
HTML touchups to the display of multipage forms.
For one-entry-per-user forms, if the user is anonymous, then a blank form is now displayed, instead of the first entry created by an anonymous user.  If a user has a cookie saved from the last time they filled in the form, then that entry will be called up instead.

From 270:
Fixed a bug where {ID} and {SEQUENCE} default values weren't working correctly (they were being stripped out by the new more flexible interpreting of {} defaults)

From 271:
Tiny naming improvement in code for building quickfilter boxes.
Multipage forms now take the title for the header row of the form, from the name of the current page.

From 272:
Added redirect_header call after the header call when saving elements, just in case the header call doesn't work (reported to fail in one case on the forums)

From 274:
Fixing a bug in the handling of uitext values in quickFilters.  Filters should work now regardless of the presence of uitext or not in the element.

From 275:
Fix to Fatal Error on line 664 of class/listOfEntriesScreen.php when using Frameworks and added FIDS don't match allFids. See http://www.freeformsolutions.ca/en/node/668

From 276:
Fixing a bug where quickfilter boxes could produce incorrect search results when a field had a long list of options, some of which contained the text of other options, ie: Baja California vs. California.

From 277:
Correcting a French translation (Autre: to Autre :)

From 278:
Fixed a bug where the disabled status of elements was not being passed back to the form rendering logic in displayForm.  This involves a change to the possible return values of displayElement.  When the $renderElement flag is false, the function now returns an array of the element object and the isDisabled value.  When the $renderElement flag is true (default), "rendered-disabled" is returned instead of "rendered".  The only known internal instances where these return values matter, have been updated to understand the changes.  External code referencing the API will need to change too.

From 279:
Added the ability to update the name of a saved view when resaving the view.

From 280:
Minor language change, related to the ability to rename saved views.

From 282:
Fixed a bug in the notification system

From 284:
Catch added for an obscure error related to frameworks and subforms (only happens when default blanks are used in a subform where the relationship between the forms is based on the user id of the user who created the entries?!)
Required asterisks now show up correctly, or don't show correctly, on the printview page in ImpressCMS.

From 285:
Added awareness of the add_own_entry permission to the Add button in the subform interface.

From 287:
Applied number formatting settings to derived values when they are displayed in a form, not just in a list.
Fixed a bug that led to elements that are part of a form but not rendered, causing an "n" to appear at the top of the form.  This was introduced with a previous change in the formdisplay.php file that fixed the disabled feature on certain kinds of elements.  (Adding list() to receive the return values of displayElements, which are not always arrays)

From 290:
Fixed a bug where grouping values for linked selectboxes were not displayed correctly.

---

Plus some special code in class/elementrenderer.php to make selectboxes and checkboxes show up properly when disabled.

git-svn-id: svn+ssh://svn.code.sf.net/p/formulize/code@291 e5624274-254a-0410-876d-c0e85fc73166