Added option to allow form elements that are displayed in columns on lists of entries screens, to not all be displayed at once, but instead be accessible through a clickable icon.
Added option to allow options in a linked selectbox to be restricted so they cannot be selected more than one time (additional options for once per user and once per group).  

Made list boxes shrink to the size of the available options, if the options are fewer than the stated size of the listbox.


git-svn-id: svn+ssh://svn.code.sf.net/p/formulize/code@530 e5624274-254a-0410-876d-c0e85fc73166