Fixed a bug that caused the relationships admin UI to screw up, when HTML is used in an element caption. (admin/relationship.php)
Tiny bit of markup cleanup (templates/admin/relationship_settings.html)

Significant improvements and optimizations and corrections to how derived value formulas are interpretted and executed within frameworks (instead of running the update once per form, it's once per framework now, and subentries within the framework are properly handled and referenced...plus element references in formulas now refer to the local entry if it's a reference in a subform to one of that subform's elements, rather than returning an array of all the subform's values).  (include/extract.php, include/functions.php)

Fixing which forms are interpretted as active when a subform entry is saved, after you've gone to the subform entry via the View buttons in the subform interface.  Improvements to the logic that finds relationships between forms, was causing the main form to be added to the screen after saving.  Now only forms that were part of the saving operation are included in the subsequent page load.  (include/formdisplay.php)

git-svn-id: svn+ssh://svn.code.sf.net/p/formulize/code@534 e5624274-254a-0410-876d-c0e85fc73166