Fixing a naming inconsistency in the dummyElement example code. (dummyElement-example.php)
Adding a space to the extra code that gets added to Formulize elements so we can track if they've changed or not, just to be on the safe side (a core fix in the standalone version should render this irrelevant, but in other versions it might matter).  (elementrenderer.php)

Adding support in extract.php for queries that filter by entry_id as a named field

Fixing a bug in the rendering of validation javascript, so that the if-statement that wraps conditional elements' validation code is only added if there is some validation code for that element in the first place.  Otherwise we return nothing.  (formdisplay.php)



git-svn-id: svn+ssh://svn.code.sf.net/p/formulize/code@668 e5624274-254a-0410-876d-c0e85fc73166