Getting all elements (even ones set not to display) to show up in the relationship-common-elements popup window. (frameCommonElements.php, frameCommonElements1.php)
Fixing bug that caused {ID} to be set incorrectly when the max entry_id did not match the autoincrement value in the table (finally!) (data.php)

Added support for the elements class to get sets of elements where the returned array can have the handle, element id or an incrementing number as the key (handle is a new option).  (elements.php)

Other formatting changes.



git-svn-id: svn+ssh://svn.code.sf.net/p/formulize/code@692 e5624274-254a-0410-876d-c0e85fc73166