A) Changed the Export function to be dynamic.Instead of  hard coding  the mechanism to export  each Table.
Function expTable($Table_Name,$Exclude_Columns = Null,$Criteria = Null,$Instructions = Null)
This function is used to export any table in the required schema, the result will be a list of insert statements for all selected rows of the table. The function depends on the INFORMATION_SCHEMA.COLUMNS to populate the attributes of the insert statements with the table field names, also it is used to identify the text fields so the values can be padded with ‘’.
Parameters
 
//Parameter Name
//Description
1
$Table_Name
Required – Table Name to Export, Prefix is padded automatically to the start of the table name
2
$Exclude_Columns
Required, User can pass here an array of the fields to execlude from export, Empty array will mean all fields
3
$Criteria
Required, Criteria (Where Condition) for selecting the required rows, Null Criteria will result in exporting the full table
4
$Instructions
Required, Specify the operations to be applied to the selected fields, currently support C to replace characters and P for Padding
 
Return: Array of Insert statements for the corresponding rows 



B)Post_Process and Check Uniqueness .I changed the functions in a way to be more dynamic and  easier to maintain.To add a new table just add the Table name to the array inside the function

C) Export and Import 2 more tables Frameworks and its links.After it imports the function checks if form 1,form 2 ,ele 1 ,ele 2 has changed or not.If yes,then it updates the change and checks if any of these elements old ids originally existed when imported.If not,then notify the user that  he is missing Form or ele.

git-svn-id: svn+ssh://svn.code.sf.net/p/formulize/code@1049 e5624274-254a-0410-876d-c0e85fc73166