Updates on Import & Export
Finished the Import and Export of the Tables.
1) Revamped the Export . Export Program simplified, code reduced by 30%.
2) The performance of the Export is much faster.
3) Unified Query Function for simple maintenance & support and.
4) Padding & Character replacing completely changed to be only dependent
on field names.

Import:
1)Changed the way the post process works .I used to carry out the post
process by using the values from the String Insert statement, now It
retrieve those data from the DB and performs the Post process on the
retrieved results.
2)Import Supports Saved_views ,Others ,Notifications, Owner group
,multipage,screen_from  ,screen ,group lists ,group filters. groupscope
settings and Advance calculations .
3)Fixed a bug that mismatched the insert statements if the Table names
are almost similar to each other .

MYSQL file
Removed the Uniq. from the Column name Desc_Form in Formulize