Adding support for nested subforms
You can show a subform as an inline form (not a row) and you can have a
subform element inside that subform, that will then be included
properly.  Should work infinitely far down, as long as all forms are in
the active relationship and tied to each other in a chain.  Also tie the
forms back to the mainform if you want to include their data in a list
of entries.