Adding support for localization of dates
date() only uses English names for months, etc. strftime will return
text based on the locale that is set.  However, it uses a different
recipe for it's format strings than the date function does!  So, we have
to add a translation function to convert the date format strings into
strftime format strings.

This is all based on taking into account the declared date formats in
the global.php file in the language folder, which is where users are
supposed to be setting this kind of thing, and are supposed to be using
the date format...I guess XOOPS was never truly localized by design!