Format numbers when displayed as text
When in elements, we are not changing anything, because it may cause
some interaction effect when 5,17 is submitted to the database, and it
gets turned into a 0 because of the comma.  We need to look into fully
localized data formats and so on.  For now, we're just showing numbers
with formatting when elements are disabled, or on printable view.