Fix for wrong language being included by host system
This is very hacky and is a response to the fact that the bootstrap
process calls the method that determines the resource mapping for the
active user.

That method call cannot lead to functions.php being included, because
the overall bootstrap is not yet complete, so including it now might
result in the wrong language file being loaded!