Optimizations for interpreting searches
Use the default LIKE operator when using a predefined quickfilter to
search multiselect elements, because the data format has a prefix on
each term, that means we have to do a pattern match, not an =, in order
to find anything.

Also, convert dates to Y-m-d format in case they are not in that format
already.