Fixing passing of source elements for export
When multiple fields were selected for displaying in the export of a
linked selectbox, we were not detecting that there were multiple and
instead were assuming they were a single item, and so they were being
nested into an array.  Now we don't stick the element list into an array
if it already is an array!