Changes to handle the rich text editor
We need a special validation javascript for when these elements are
required.

We set a global flag to indicate to the drawJavascript function that we
need to ask the FCK api whether the contents have changed, so we can set
formulizechanged properly.

We strip tags from the saved contents when displaying them in a list,
and limit them to the first 100 characters.  This should make the list
look OK and not bent out of shape due to formatting and spacing, and
amount of text entered into the rich text editor.

The FCK editor API was used throughout this, which would make supporting
another editor a non-automatic process.  It was actually pretty straight
forward to handle all the cases we needed to cover, using the API, so
yay.

This should be cross platform with XOOPS.