Fix for 'other' boxes not being editable
'Other' boxes were included inside the <label> for the radio button, so when you clicked on them in at least some browsers, they lost focus because the radio button became clicked instead.  We already have javascript onfocus for the 'other' boxes that selects the radio button anyway, and this fix makes it possible for the focus to be retained in the box so people can type.