Require module admin for Configure screen link
Registered Users actually have Edit Form permission on all new forms, as a safeguard against users who have rights to create a form, not being able to edit their newly created forms!  But you need module admin rights to get into the backend at all, so we should be checking that perm as well as the right to edit the form.