Bug fix in previous commit
We need to gather $value from $prevEntry before we can determine if it
is empty or not.

if(!$value) at the top of that block refers to whether a value came back
for this element, from the form submission that led to this page, or
not.

Also amended the logic so required fields are treated this way also.  A
required field, by definition, should not be empty in the database, so
if something has happened to make it empty, we'll ignore that blank
value and use the default instead.