Fix bug + set current user as cloned entry owner
Previously we had been perpetuating the ownership info of the creator of
the original entry.  But now we use the creator of the cloned entry,
which we are now correctly letting default to the currently active user.
Previously we were passing $xoopsUser which was causing a bug and cloned
entries were owned by user 1 as a result!