Reverting 3628e5fc58e4ebd33ebeeaa38d5e2b32fad6a152
There is already an exit line above for when there is not entry id identified. The CFCC code must lack that, or else the error users encountered would have resulted in a halt before formulize_writeEntry, and the error text in the exit line would have been printed on screen.

The more serious exit is a little brutal, but makes it dead clear what's happened, and highlights a data integrity issue, which ought to be dealt with right away through an adjustment to the configuration of the site somehow.  The intention is that this would be easily found during the setup of a feature.