treat one-to-one elements as conditionals
In addition to tracking the elements that have conditions, also track
elements that are part of one-to-one relationships, and include them in
the same process as conditional elements, so that they can swap in/out
depending on the key value selected in the joining element that ties
them back to the main form.  This necessitated changes to how the
conditional logic worked, slightly.