Add flag to pres. quicksearches when load reports
Instead of resetting all columns to match the values saved in the
report, we now keep any values the user has typed in/selected.  If the
report has a diff. value for a given column then the value from the
report will replace the user value. Other user values will remain in
place and not be neutered.