Moved gathering of hidden elements
We gather "forcehidden" elements for inclusion in a form now, based on
the elements that are actually displayed to the user, ie: any element in
the form that has forcehidden set, that is not included in the available
elements on the screen for the user to interact with, will now be
treated as a element-for-inclusion-as-hidden.  Previously, only elements
where the display-to-these-groups choice excluded the user, would be
used.  Now if the user doesn't see the element for any reason, it is
used.