Adding lastentry to all forms
When there is an entry number, we add lastentry to the form.  The
lastentry field was originally only in the "go back" form triggered by
the alldone button.  It is used to indicate that we're coming back from
a form, and it will prevent pagenumbers from being regenerated in a
list.  Without including it directly in the form, then a multipage form
that returns to the list without a thank you page, will not have
lastentry and so it will always reset the page number to 1.