Attempting to save elements within sub-ui-accordions
1. The elements save code has been moved into elements_save.php since it
is being used by both forms and screen now
2. Each sub-ui-accordion.html will contain a hidden tag and will be
called when the save button is click. This will then populate the
element order and drawer number for this accordion.
3. Within screen_multipage_pages_save.php we use a regex expression to
go through all of the possible accordions and call save_elements() for
each one of them.

Note: It currently runs the code at this point, but for screen, the
elements order is still not saved.