Treat typed val as new val if box loses focus early
This is an incomplete fix! It will lead to duplicates if the user types
a value that matches an existing one and then leaves the box early
before the autocomplete UI shows up (typed value will be saved as if it
were new, not picking up the id of the existing value).

Also, JS minification might need to be redone. I did it manually.