Added case 'masquerade'
$xoopsUser contains all user permissions/information. I figured out how
to create new user instance to replace $xoopsUser, but the variable does
not persist between webpages.
Will next try to change the session instead of $xoopsUser. Will need to
figure out how a session is created or maintained.
After a successful masquerade, I plan to redirect the user back to the
main page (this part is ready).