Added to template, created custom_edit_check methods, created table entry in sql file
On the html template, I created a new code entry space where the user
can type the value for $custom_edit_check. In forms.php, I mirrored all
of on_before_save()'s methods, and initialized the $custom_edit_check
variable at the top. around line 530, I added to the SQL statements to
insert the custom_edit_check code into the database, and finally in the
mysql.sql file, I added a line to create the column for
custom_edit_check. From my error log, the variable is being populated
with what I type into the input box. The code is also being cached.

TO DO: Patch the database so that the column can be inserted into, check
that the user has the required variables to set  up their condition, and
make sure the check is run at the appropriate time.