Finishing pieces of the code for desired output of customEditCheck() Function
It appears a second return value was necessary in order to have the
customEditCheck() function operate properly. upon looking over
everything and adding this, the test case below was implemented without
error:
"if($entry_id % 2 == 0){ // even values
$allow_editing = 1;
}else{
$allow_editing = 0;
}"
I was able to edit the even numbered entries, but not the odd, as
desired.