Alter CSS to prevent Overlapping
The formulize-admin.css file contained one div (homepage-description)
which was to be limited in its width. The other div which contained the
toolbar, had no name or space in the stylesheet, as it was hard-coded
into ui-accordion.html. I added a width limit to this as well, and now
the elements do not overlap.