Further CSS Fixes, Toolbar placement
In home .php, I added the 'add a form' button to the toolbar, and
removed it from the 'home_sections.html' template.
removed previous edit to ui-accordion, and made a change to the margin
in '.homepage-description', (formulize-admin.css).
Finally, as a precaution to eliminate any further overlapping of
elements on the admin page, I added a minimum width of 803 pixels
(measured least pixels required with a web dev. ruler) to the
'#xo-canvas-content' div.