Button Implementation for Deleting Temporary files.
A New script was made in the Formulize root folder called
'deleteTempCache.php', which runs a small script to delete the cached
template files in a for each loop.
The button was made in home.php to be in line with the right-hand menu.
In ui,.html, the JavaScript was appended to include a click function for
the above button, which (only when confirmed), forms an AJAX query for
the newly created script, which is executed to delete cached files.
The min-width of the xo-canvas was also increased to accommodate the
change in size of adding in the new button to prevent overlap.