Requested changes to new Delete Cached Templates button
The code from the original 'deleteTempCache.php' file was moved into the
'formulize_xhr_responder.php' handler, and the "DOCUMENT_ROOT" call was
changed  to use the XOOPS_ROOT_PATH, for consistency throughout servers.
In home.php, the button was moved into home.html, into the same div with
the same positioning as the other entries.
Finally in ui.html, the code was altered to use <{$xoops_url}> to
provide a consistent base in any server.