JQuery click functions for Check Boxes
I have placed in this everything we have discussed on the pull request.
Upon clicking the checkbox, it checks to see if there is a value in the
input box, AND that the checkbox was checked (does nothing if
unchecked).
If checked, it will focus on the box and set the value initially to
Zero.