Fix a miss matched element type
Due to this miss matched type, when the file upload box is rendered, it
will be treated as xoopsFormLabel rather than XoopsFormFile object. When
it comes to xoopsFormLabelobject, it will not contain
onchange="javascript:formulizechanged=1"  for file upload box, but this
part is the key to tell if there is a value changed.

I think this will fix the problem.