Adding special cleaning function for ops
Previously, we used sanitization functions to clean the possible ops,
but since < > are valid chars in ops, this was leading to bizarre
outcomes as html special chars appeared as part of filter ops and
eventually got into SQL if you resaved enough times, and basically bad
stuff happened.

This way, we simply filter out ops based on whether they match the
allowed strings or not. In any systems affected by this bug, we may need
to recreate the filter conditions, because it is possible in rare cases
that the op will be nerfed by the new function. This will be visible in
the admin UI so we have a clue that we need to recreate the filter
condition.