Add support for curlybracket terms in other forms
Curly bracket terms may be tied to a form outside the current form,
which is useful in some situations when prepop subform elements are
filtering options based on values in a parent form. But when rendering
the options for such a linked selectbox, we must ignore the filtering,
and hope that the field is disabled probably! or else all options will
show up.