Added support for translated options in elements
If options were translated, then we would lose the equivalency of what
the user chose in the past, with that the options are now. To
compensate, we check the English translation values of the selections as
a safety net, if there's been no matches so far. This will result in
legacy values behaving in the UI as if they are the new equivalent
values based on the English translation. Strong assumption is that the
original language deployed was English!!