Adding support for custom scope filter
Writing custom code in the prefs will result in unique rules for how the
scope is interpretted. An array of group ids will be passed in, in most
cases, and for an individual site if the scope can be simplified to user
ids, then returning a string of them will speed up performance.
Dramatically on a large database.