Adding support for multiple blanks in subsub forms
If there are blanks in a sub sub form, then you can end up with multiple
new blanks that need to be synched with different source entries. This
code makes the assumption that there is one blank per source. This WILL
NOT WORK if there are multiple blanks per source, we would need
something smarter then.