Entry Id 0 valid for displaying but not saving
In the case of dynamic one to one entries (TAs/TAships) where the
linking value has been removed/neutered, we need to generate an empty
field, but since the linking field is empty we can't actually save any
of the data (and since elements can't be conditional based on elements
in another one to one form, we have to still show the element). So it
needs to have a dummy value in some way that won't be saved. Setting
entry id explicitly to zero accomplishes this.