Added smarts to ignore { } terms that are invalid
If a { } term is not set in $_GET and is not a valid handle in the
target form, then ignore that part of the filter entirely (otherwise
filter would always return no results). This is appropriate and useful
for cases where a single element needs to be filtered differently in
different situations, relying on different URL params in each case. All
applicable filter terms can be added to the element's definition, but
they will be selectively ignored based on context.