Smarter handling of BR tags
So other tags in a paragraph with breaks will still be interpretted. An
awful hack though, but I don't see a way around it because reading the
DOM makes BRs into their own element, and an XML document with an
analagous structure is going to treat the breaks as independent
elements, which is going to introduce extra spaces around them.