Further enhancing searches when pubfilters in use
This will allow the master view of a form/framework to preserve hidden
column searches that might be in effect for purposes of bring picked up
by a published filter. Otherwise, they would be lost on next pageload,
because on the master page, there would be no instantiation of the
published filter through a template. Important for admin side usability.