Derived values update when mainform isn't written
If an entry in a form is written, and it's in a screen where there's a
relationship, and the mainform in the relationship isn't written, then
we don't flag the form that was written as having been updated, even
though a corresponding entry was found in the relationship. Instead, the
form that was written will be updated as part of the entries not in the
relationship. HOWEVER, this means that the derived values in the
non-mainform need to function independent of the relationship, since
they are being updated outside the relationship, even though the screen
in effect has a relationship that includes both forms. Mainform isn't
written, so relationship is not the auspices under which the updating
takes place.