Adding support for passed in comparison values
Manually set a GLOBAL to specify what a given element should be compared
against (instead of the value in the DB). Useful only in cases where you
need to compare to something that is not going to be in the dataset or
in the data table you're talking with, so pass in a subquery, for
example.