Update the database column to varchar
In the file 'formindex.php', a new $sql entry was created updating the
column 'fl_key1' in the database table 'formulize_framework_links' from a
smallint to a varchar(10). This is because the new values passed from 'relationship_settings.html' from the new selection dropdown entry values sent  to the database are strings, and must be stored accordingly. Then in 'formindex.php' the 4 variables (beginning with $checkthistable) that specify a patch must be completed are updated to reflect the new changes.