The working icons and some prototype
all png files are the first drafts of the icons. The file initialMaker is a php script which will be incorperated into home.php so that the initials of the file name can be returned through javascript. testSlider is a very basic layout of how the sliders should behave. All files are in a new folder called UI Testing so as it should not interfere with the rest of the code for the time being.