Adding prepop feature for subforms (#129)
* Adding prepop feature for subforms

* test commit to confirm passing after merge with master

* corrected html ID for num subform blanks

* 018 test for prepopulation form entries via dropdown list.

* 018 adjusted xpath

* updated xpath after fresh DB wipe

* Refactoring to always write defaults when needed

* Fixing mixup of element ids/handles

obscure, rare, nasty

* Adding ignoring of .sass-cache files

* Add support for curlybracket terms in other forms

Curly bracket terms may be tied to a form outside the current form,
which is useful in some situations when prepop subform elements are
filtering options based on values in a parent form. But when rendering
the options for such a linked selectbox, we must ignore the filtering,
and hope that the field is disabled probably! or else all options will
show up.

* Set prepop element conditions

If an element that governs prepopulating of a subform has conditions,
then use those conditions when creating the new entries.

* Allow conditions on subforms to affect prepops

If a subform has conditions on which elements are to be included, those
will affect the generation of prepop entries.

* Fixing typo