Add field to "formulize" (elements) table.
Add the 'ele_foreign_key_element_link' field to the "formulize" (aka elements)
table that is of type TINYINT(1) representing a boolean. The field
'ele_foreign_key_element_link' represents whether the current element is
the linked element between the forms. The particular linked element should be
invisible to the user. The addition of the field to the "formulize"
table is in the "formindex.php" file. Also, the "mysql.sql"
file is changed to reflect the field for new installs. Delete a
previously wrongfully placed field in the "formulize_framework_links"
table and update the "mysql.sql" file to reflect this change for new
installs.