Add method in form handler to check if invisible linked element exists.
In the "forms.php" file, add a method to check all elements in the
target form for a hidden element used to link two forms. Use this
method in the "relationships.php" file to decide whether or not to
create a new hidden element, and join the two forms using this element.

Also fixed a small code bug in the "elements.php" file.