Use database values to evaluate display conditions
If an element uses ui text, we don't want to do a conversion. We need to
base the comparison on the value coming back from the getData/prepValues
process (and the asynchronous override in the xhr_responder if it is a
conditional element). This code was probably put in place back when
prepValues was converting to use uitext in all cases (as in Dara), but
now that we only use the uitext in the form rendering itself, never in
the getData results, we need to not to the conversion here.