Create new element and insert it into 'formulize' table (unfinished)
A new empty element is created using the element handler in
'elements.php'. Once the element is created, values are assigned to the
objects properties. The new element is then inserted into the
'formulize' table with other elements.

Problems: - Currently, the element is being created and the values are
being assigned, however it is refusing to be added into the 'formulize'
table via the 'insert()' method on the elements handler.