Hide linking elements from form UI
Change the SQL statements in the "get" and "&getobjects" methods in the
element handler to ignore the newly created linking elements by adding a
check if the "ele_foreign_key_element" is equal to 0. Also, fixed a bug
that prevented the page from loading when the form UI for a linked page was called and caused the page to fail to load. This happened because the newly created element didn't have a type. The type is now specified to "text".